Refactor slope to eliminate use of GraphX

+++ mrgeo-core/src/main/scala/org/mrgeo/spark/job/MrGeoDriver.scala
@@ -36,7 +36,8 @@ abstract class MrGeoDriver extends Logging {
 
   def setup(job: JobArguments): Boolean
 
-  def run(name:String, driver:String = this.getClass.getName, args:Map[String, String] = Map[String, String](), hadoopConf:Configuration) = {
+  def run(name:String, driver:String = this.getClass.getName, args:Map[String, String] = Map[String, String](),
+      hadoopConf:Configuration) = {
     val job = new JobArguments()
 
     job.driverClass = driver
@@ -212,7 +213,7 @@ abstract class MrGeoDriver extends Logging {
 
   }
 
-  private def calculateYarnResources():(Int, Int, Int) = {
+  private def calculateYarnResources():(Int, Int, Long) = {
 
     val cl = getClass.getClassLoader
 
@@ -259,7 +260,7 @@ abstract class MrGeoDriver extends Logging {
 
     stop.invoke(yc)
 
-    (cores, nodes, memory.toInt)
+    (cores, nodes, memory)
   }
 
   private def configureYarnMemory(job:JobArguments, memory:Long, minmemory:Long,
